Split dashboard to components, Fix missing validation, add logs
