+++ server/src/domain/game/Room.ts
@@ -7,6 +7,7 @@ import {
 } from '@quasm/common';
 import { UUID } from 'crypto';
 
+import { logger } from '@/infrastructure/logger/Logger';
 import { IRoomRepository } from '@/repositories/room/IRoomRepository';
 
 import { Character, CharacterDetails } from './Character';
@@ -19,8 +20,9 @@ export class RoomSettings {
     constructor(
         public roomName: string = '',
         public maxPlayerCount: number = 1
-    ) {}
+    ) { }
 }
+
 export class Room {
     private gameMaster?: UUID;
     private characters: Character[] = [];
@@ -32,7 +34,52 @@ export class Room {
         // private storyChunks: StoryChunk[] = [];
         // private broadcast: Chat;
         // private chats: Map<(UUID, UUID), Chat>; ???
-    ) {}
+    ) {
+        this.validateName(this.getName());
+        this.validateMaxPlayerCount(this.getMaxPlayerCount());
+    }
+
+    validateName(name: string) {
+        if (name.length <= 0 || name.length > MAX_ROOM_NAME_LENGTH) {
+            throw new QuasmError(
+                QuasmComponent.ROOM,
+                400,
+                ErrorCode.MaxRoomName,
+                'Incorrect name'
+            );
+        }
+    }
+
+    validateMaxPlayerCount(count: number) {
+        if (count < 2 || count > MAX_ROOM_PLAYERS) {
+            throw new QuasmError(
+                QuasmComponent.ROOM,
+                400,
+                ErrorCode.IncorrectMaxPlayers,
+                'Incorrect max players'
+            );
+        }
+    }
+
+    validateCharacter(character: CharacterDetails) {
+        if (this.characters.length >= this.roomSettings.maxPlayerCount) {
+            throw new QuasmError(
+                QuasmComponent.ROOM,
+                500,
+                ErrorCode.MaxPlayersExceeded,
+                'Too many characters restored'
+            );
+        }
+
+        if (this.characters.find(({ userID }) => userID === character.userID)) {
+            throw new QuasmError(
+                QuasmComponent.ROOM,
+                400,
+                ErrorCode.UserExists,
+                `Room: ${this.id} User: ${character.userID}`
+            );
+        }
+    }
 
     /**
      * Helper method for reassigning character to the Room, for adding new character to the room use {@link addCharacter}
@@ -47,8 +94,17 @@ export class Room {
             );
         }
 
+        logger.info(
+            QuasmComponent.ROOM,
+            `Restoring character: ${character.id}`
+        );
+
         this.characters.push(character);
         if (character.isGameMaster) {
+            logger.info(
+                QuasmComponent.ROOM,
+                `Game master restored: ${character.id}`
+            );
             this.gameMaster = character.id;
         }
     }
@@ -68,14 +124,7 @@ export class Room {
     }
 
     async addCharacter(characterDetails: CharacterDetails) {
-        if (this.characters.length >= this.roomSettings.maxPlayerCount) {
-            throw new QuasmError(
-                QuasmComponent.ROOM,
-                500,
-                ErrorCode.MaxPlayersExceeded,
-                'Too many characters restored'
-            );
-        }
+        this.validateCharacter(characterDetails);
 
         const character = await this.roomRepository.addCharacter(
             this.id,
@@ -108,14 +157,7 @@ export class Room {
     }
 
     async setName(newName: string) {
-        if (newName.length <= 0 || newName.length > MAX_ROOM_NAME_LENGTH) {
-            throw new QuasmError(
-                QuasmComponent.ROOM,
-                400,
-                ErrorCode.MaxRoomName,
-                'Incorrect name'
-            );
-        }
+        this.validateName(newName);
 
         await this.roomRepository.updateRoom(this.id, {
             ...this.roomSettings,
@@ -130,14 +172,7 @@ export class Room {
     }
 
     async setMaxPlayerCount(newMaxPlayerCount: number) {
-        if (newMaxPlayerCount < 2 || newMaxPlayerCount > MAX_ROOM_PLAYERS) {
-            throw new QuasmError(
-                QuasmComponent.ROOM,
-                400,
-                ErrorCode.IncorrectMaxPlayers,
-                'Incorrect max players'
-            );
-        }
+        this.validateMaxPlayerCount(newMaxPlayerCount);
 
         await this.roomRepository.updateRoom(this.id, {
             ...this.roomSettings,
